Try to generate constant values via jit.
